Find dropped messages lines from rsyslog and suggest a higher burst level
Rsyslog will drop messages from processes that are reporting too many messages
within a given interval.  If this happens, information can be lost.  This may
mean that you need to rein in a runaway process, or deal with errors; or it
may mean that you need to raise the burst rate so that you can read more
errors :-)  In order to be a bit more useful, we report:

* What processes were being dropped.
* What the current rsyslog rate limiting interval and burst limits are.
* What the new configuration parameter should be to avoid this problem.